% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2blog-package.R
\docType{package}
\name{add2blog-package}
\alias{add2blog}
\alias{add2blog-package}
\title{add2blog: blog website add-in}
\description{
add2blog is R package to create some materials for a blog website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JiaxiangBU/add2blog}
  \item Report bugs at \url{https://github.com/JiaxiangBU/add2blog/issues}
}

}
\author{
\strong{Maintainer}: Jiaxiang Li \email{alex.lijiaxiang@foxmail.com}

}
\keyword{internal}
