% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2blog-package.R
\docType{package}
\name{add2blog}
\alias{add2blog}
\alias{add2blog-package}
\title{add2blog: help your blog site}
\description{
The goal of add2blog is to help you build some materials like wall paper for your blog website !
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JiaxiangBU/add2blog}
  \item Report bugs at \url{https://github.com/JiaxiangBU/add2blog/issues}
}

}
\author{
\strong{Maintainer}: Jiaxiang Li \email{alex.lijiaxiang@foxmail.com}

}
\keyword{internal}
