% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_download.R
\name{badge_download}
\alias{badge_download}
\title{Get a badge for the download count of your package}
\usage{
badge_download(your_github_name = "JiaxiangBU",
  your_github_repo = "add2prep", type = c("total"))
}
\arguments{
\item{your_github_name}{character}

\item{your_github_repo}{character}

\item{type}{character copy it.}
}
\value{
character
}
\description{
Input your github name and repo name,
this function will feedback the download count of your package.
}
\examples{
add2blog::badge_download(
your_github_name = 'JiaxiangBU'
,your_github_repo = 'add2prep'
,type = 'total')
}
