% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogdown_meta.R
\name{blogdown_meta}
\alias{blogdown_meta}
\title{Get Metadata from all your blogs built by blogdown}
\usage{
blogdown_meta(dir = ".")
}
\arguments{
\item{dir}{Character. The directory contains all posts you want to analyze.}
}
\value{
data.frame
}
\description{
This function is inspired by function `count_yaml` in the package `blogdown`.
}
\details{
When you post a lot for your blog site built by `blogdown`, you may want to merge, edit metadata of some of them.
The function `blogdown_meta` go deeper in the `yaml` contents of each `.Rmd` and
build a friendly readable `data.frame` to analyze.
}
\examples{
blogdown_meta(dir = '.')
}
\author{
Jiaxiang Li
}
