% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogdown_meta_data.R
\docType{data}
\name{blogdown_meta_data}
\alias{blogdown_meta_data}
\title{blogdown_meta_data blodgown meta data}
\format{A data frame of 287 rows and 6 columns
\describe{
 \item{dir}{Character}
 \item{author}{Character}
 \item{categories}{Character}
 \item{categories}{Character}
 \item{date}{Character}
 \item{tags}{Character}
 \item{title}{Character}
}}
\source{
https://jiaxiangli.netlify.com
}
\usage{
blogdown_meta_data
}
\description{
A dataset containing the real blodgown meta data
}
\keyword{datasets}
