% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_fun.R
\name{graph_fun}
\alias{graph_fun}
\title{Build a wallpaper from your name}
\usage{
graph_fun(matrix, fontsize = 2.5)
}
\arguments{
\item{matrix}{input}

\item{fontsize}{font size}
}
\value{
ggplot
}
\description{
Input a matrix and build a wallpaper.
This is a input for function `graph_fun` to create a wallpaper.
}
\examples{
matrix_me <-
  add2blog::matrix_name('jiaxiangli',n = 10)
add2blog::graph_fun(matrix_me,fontsize=3)
}
