% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png2gif.R
\name{png2gif}
\alias{png2gif}
\title{Convert pictures to gif}
\usage{
png2gif(path = ".", input_like = ".png", output_name = "tmp.gif",
  size = "1000x1000", transfer_time = 1)
}
\arguments{
\item{path}{Character. The directory contains the pictures to convert.}

\item{input_like}{Character. The file names of the pictures like}

\item{output_name}{Character. The output file name}

\item{size}{Character. n x n.}

\item{transfer_time}{Integer. n seconds.}
}
\value{
gif
}
\description{
Convert screenshots or other pictures into a gif.
}
\examples{
png2gif(
path = file.path('../','inst')
,input_like = 'reprex'
,output_name = 'reprex201811061210.gif'
)
}
\author{
Jiaxiang Li
}
