% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2blog-package.R
\docType{package}
\name{add2blog}
\alias{add2blog}
\alias{add2blog-package}
\title{add2blog: help your blog site}
\description{
The goal of add2blog is to help you build some materials like wall paper for your blog website !
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JiaxiangBU/add2blog}
  \item Report bugs at \url{https://github.com/JiaxiangBU/add2blog/issues}
}

}
\author{
\strong{Maintainer}: Jiaxiang Li \email{alex.lijiaxiang@foxmail.com}

Other contributors:
\itemize{
  \item Yihui Xie [contributor]
  \item Beilei Bian [contributor]
  \item Forest Fang [contributor]
  \item Garrick Aden-Buie [contributor]
  \item Hiroaki Yutani [contributor]
  \item Ian Lyttle [contributor]
  \item JJ Allaire [contributor]
  \item Kevin Ushey [contributor]
  \item Leonardo Collado-Torres [contributor]
  \item Xianying Tan [contributor]
  \item  RStudio Inc [copyright holder]
}

}
\keyword{internal}
