% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_name.R
\name{matrix_name}
\alias{matrix_name}
\title{Randomize your name in a matrix}
\usage{
matrix_name(x, n = 10)
}
\arguments{
\item{x}{character your name}

\item{n}{size n by n}
}
\value{
martix n by n
}
\description{
Capitalize each letter of your name and randomly put in a matrix.
This is a input for function `graph_fun` to create a wallpaper.
}
\examples{
matrix_me <-
  add2blog::matrix_name('jiaxiangli',n = 10)
}
